refactor; move quickaction component to common globals, structure config page
