+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Configuration/BasicInformation/graphql/client/update-app.generated.ts
