+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Configuration/BasicInformation/index.tsx
@@ -16,9 +16,9 @@ import * as yup from "yup";
 import {
   FetchAppMetadataDocument,
   FetchAppMetadataQuery,
-} from "../../graphql/client/fetch-app-metadata.generated";
-import { viewModeAtom } from "../../layout/ImagesProvider";
-import { RemainingCharacters } from "../../PageComponents/RemainingCharacters";
+} from "../graphql/client/fetch-app-metadata.generated";
+import { viewModeAtom } from "../layout/ImagesProvider";
+import { RemainingCharacters } from "../PageComponents/RemainingCharacters";
 import { useUpdateAppInfoMutation } from "./graphql/client/update-app.generated";
 
 const schema = yup.object({
@@ -186,6 +186,8 @@ export const BasicInformation = (props: {
             <Typography variant={TYPOGRAPHY.M3}>Save Changes</Typography>
           </DecoratedButton>
         </form>
+
+        <div></div>
       </div>
     </div>
   );
