+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Configuration/graphql/client/fetch-team-name.generated.ts
