+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Configuration/page.tsx
@@ -2,11 +2,11 @@
 import { SizingWrapper } from "@/components/SizingWrapper";
 import Error from "next/error";
 import Skeleton from "react-loading-skeleton";
-import { useFetchAppMetadataQuery } from "../graphql/client/fetch-app-metadata.generated";
-import { AppTopBar } from "../PageComponents/AppTopBar";
-import { FormSkeleton } from "../PageComponents/AppTopBar/FormSkeleton";
 import { BasicInformation } from "./BasicInformation";
+import { useFetchAppMetadataQuery } from "./graphql/client/fetch-app-metadata.generated";
 import { useFetchTeamNameQuery } from "./graphql/client/fetch-team-name.generated";
+import { AppTopBar } from "./PageComponents/AppTopBar";
+import { FormSkeleton } from "./PageComponents/AppTopBar/FormSkeleton";
 
 type AppProfilePageProps = {
   params: Record<string, string> | null | undefined;
